Add bytecode compiler + VM with native opcodes and 49 tests
Bytecode compiler and stack-based VM for FRUN command, targeting
faster execution on RP2040/RP2350. Includes:

Compiler (bc_compiler*.c):
- Two-pass compilation: label/sub/function scan then code emission
- Native opcodes for control flow (IF/FOR/DO/WHILE/SELECT CASE),
  subroutines, GOTO/GOSUB, DIM, PRINT, DATA/READ/RESTORE
- Native string functions: LEN, LEFT$, RIGHT$, MID$, UCASE$, LCASE$,
  INSTR (2 and 3-arg), CHR$, ASC, VAL, STR$, HEX$, OCT$, BIN$,
  SPACE$, STRING$, INKEY$
- Native math functions: SIN, COS, TAN, ATN, ASIN, ACOS, ATAN2,
  SQR, LOG, EXP, ABS, SGN, INT, FIX, CINT, RAD, DEG, PI, MAX, MIN, RND
- Native statements: INC, CONST, RANDOMIZE, ERROR, CLEAR
- Bridge fallback for all other commands/functions

VM (bc_vm.c):
- Computed-goto dispatch for fast opcode execution
- Stack-based with separate int/float/string value stacks
- Local variable frames for SUB/FUNCTION scope
- Rotating string temp buffers to avoid allocation

Host test build (host/):
- Native macOS build for off-device testing
- 49 compare-mode tests verifying interpreter/VM output match
- Covers: arithmetic, strings, control flow, recursion, arrays,
  type coercion, DATA/READ, bridge functions, edge cases

+++ MMBasic.c
@@ -458,7 +458,7 @@ int   MIPS16 PrepareProgramExt(unsigned char *p, int i, unsigned char **CFunPtr,
     }
     while(*p == 0) p++;                                             // the end of the program can have multiple zeros
     p++;                                                            // step over the terminating 0xff
-    *CFunPtr = (unsigned char *)(((unsigned int)p + 0b11) & ~0b11); // CFunction flash (if it exists) starts on the next word address after the program in flash
+    *CFunPtr = (unsigned char *)(((uintptr_t)p + 0b11) & ~(uintptr_t)0b11); // CFunction flash (if it exists) starts on the next word address after the program in flash
     if(i < MAXSUBFUN) subfun[i] = NULL;
     CurrentLinePtr = NULL;
     // now, step through the CFunction area looking for fonts to add to the font table
